requirements.txt
streamlit
requests
pandas
websocket-client
google-cloud-storage
